a little reorg
