+++ pir/shard_test.go
@@ -1,7 +1,6 @@
 package pir
 
 import (
-	"fmt"
 	"math/rand"
 	"testing"
 )
@@ -105,29 +104,6 @@ func HelperTestShardRead(t *testing.T, shard Shard) {
 			t.Fatalf("response2 is incorrect. byte %d was %d, not '%d'\n", i, res[i], expected)
 		}
 	}
-
-}
-
-func TestShardCPUReadv0(t *testing.T) {
-	fmt.Printf("TestShardCPUReadv0: ...\n")
-	shard, err := NewShardCPU("shardcpuv0", TestDepth*TestMessageSize, generateData(TestNumMessages*TestMessageSize), 0)
-	if err != nil {
-		t.Fatalf("cannot create new ShardCPU v0: error=%v\n", err)
-	}
-	HelperTestShardRead(t, shard)
-	afterEach(t, shard, nil)
-	fmt.Printf("... done \n")
-}
-
-func TestShardCPUReadv1(t *testing.T) {
-	fmt.Printf("TestShardCPUReadv1: ...\n")
-	shard, err := NewShardCPU("shardcpuv1", TestDepth*TestMessageSize, generateData(TestNumMessages*TestMessageSize), 1)
-	if err != nil {
-		t.Fatalf("cannot create new ShardCPU v1: error=%v\n", err)
-	}
-	HelperTestShardRead(t, shard)
-	afterEach(t, shard, nil)
-	fmt.Printf("... done \n")
 }
 
 func HelperBenchmarkShardRead(b *testing.B, shard Shard, batchSize int) {
@@ -150,23 +126,4 @@ func HelperBenchmarkShardRead(b *testing.B, shard Shard, batchSize int) {
 		}
 	}
 	b.StopTimer()
-
-}
-
-func BenchmarkShardCPUReadv0(b *testing.B) {
-	shard, err := NewShardCPU("shardcpuv0", BenchDepth*BenchMessageSize, generateData(BenchNumMessages*BenchMessageSize), 0)
-	if err != nil {
-		b.Fatalf("cannot create new ShardCPU v0: error=%v\n", err)
-	}
-	HelperBenchmarkShardRead(b, shard, BenchBatchSize)
-	afterEach(b, shard, nil)
-}
-
-func BenchmarkShardCPUReadv1(b *testing.B) {
-	shard, err := NewShardCPU("shardcpuv1", BenchDepth*BenchMessageSize, generateData(BenchNumMessages*BenchMessageSize), 1)
-	if err != nil {
-		b.Fatalf("cannot create new ShardCPU v1: error=%v\n", err)
-	}
-	HelperBenchmarkShardRead(b, shard, BenchBatchSize)
-	afterEach(b, shard, nil)
 }
